% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_netcdfs.R
\name{calculate_weighted_mean}
\alias{calculate_weighted_mean}
\title{Calculate the weighted means}
\usage{
calculate_weighted_mean(dt, intermed_dir, output_dir)
}
\arguments{
\item{dt}{a data.table of CMIP information with the area file to process.}

\item{intermed_dir}{the directory path indicating where to save intermeidate output files to.}

\item{output_dir}{the directory location indicating where to save the final csv output to.}
}
\value{
A data.table of the datetime, year, and month.
}
\description{
\code{calculate_weighted_mean} Use gridded netcdfs to calculate the weighted area for some
fld mean.
}
