% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.distgrid.R
\name{cdo.distgrid}
\alias{cdo.distgrid}
\title{R Interface for 'cdo.sellonlatbox' and 'cdo.distgrid'}
\usage{
cdo.distgrid(x, y, ..., dim = c(1, 1), return.raster = TRUE)
}
\arguments{
\item{x, y}{BLA BLA}
}
\description{
R Interface for 'cdo.sellonlatbox' and 'cdo.distgrid'
}
\examples{

\dontrun{

	ncname <- "chirps-v2.0.2005.days_p05.nc"
 url <- "ftp://ftp.chg.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/global_daily/netcdf/p05"
 temp <- rasterTmFile()
 download.file(from=paste(url,ncname,sep="/"),to=temp)
 x <- paste(temp,ncname,sep="/")



### Africa Extent: extent      : -18.1625, 54.5375, -34.8375, 37.5625  (xmin, xmax, ymin, ymax)
###y <- extent(c(-17.6608, -5.727799, 10.17488, 18.62197)) ## Extent senegal
 y <- extent(c(-18.1625,54.5375,-34.8375,37.5625))

 prec_afr <- cdo.distgrid(x=x,y=y,dim=c(10,10),parallel=TRUE)	


}
}

