% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.cdo.remapnn.conf.R
\name{cdo.remapnn.conf}
\alias{cdo.remapnn.conf}
\title{It applies 'cdo.remapnn' to the netcdf file listed in the CONFIGURATION file}
\usage{
cdo.remapnn.conf(file = system.file("configuration_template/CONFIGURATION.R",
  package = "nccdosmart"), outdir = "./temporary_nccdosmart", files = NULL)
}
\arguments{
\item{file}{CONFIGURATION file}

\item{outdir}{output directory for newly created NetCDF}

\item{files}{NetCDF data file if not set through the CONFIGURATION file}

\item{...}{further arguments for \code{\link{cdo.remapnn}}}
}
\description{
It applies 'cdo.remapnn' to the netcdf file listed in the CONFIGURATION file
}
\note{
This function calls \code{cdo seldate}.
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\examples{
out <- cdo.remapnn.conf()
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{cdo.seldate.conf}},\code{\link{system}},\code{\link{cdo.remapnn.conf}}
}

