% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.cdo.remapnn.R
\name{cdo.remapnn}
\alias{cdo.remapnn}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.remapnn(coarse_nc, fine_nc, outfile)
}
\arguments{
\item{coarse_nc}{coarse input netcdf}

\item{fine_nc}{fine  input netcdf}

\item{outfile}{output netcdf}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo remapnn}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

