% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.cdo.conf.R
\name{cdo.seldate.conf}
\alias{cdo.seldate.conf}
\title{It applies 'cdo.seldate' to the netcdf file listed in the CONFIGURATION file}
\usage{
cdo.seldate.conf(file = system.file("configuration_template/CONFIGURATION.R",
  package = "nccdosmart"), outdir = "./temporary_nccdosmart")
}
\arguments{
\item{file}{CONFIGURATION file}

\item{outdir}{output directory for newly created NetCDF}

\item{...}{further arguments for \code{\link{cdo.seldate}}}
}
\description{
It applies 'cdo.seldate' to the netcdf file listed in the CONFIGURATION file
}
\note{
This function calls \code{cdo seldate}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\examples{

 out <- cdo.seldate.conf()


##(file="/Users/ecor/workspace_eclipse/MASTERCLASS/EMC15ENERGY/CONFIGURATION",
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{cdo.seldate}},\code{\link{system}},\url{http://pcmdi9.llnl.gov/esgf-web-fe/live#}
}

