% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.seldate.R
\name{cdo.seldate}
\alias{cdo.seldate}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.seldate(infile, outfile, start = "1997-05-01", end = "1997-05-31", ...)
}
\arguments{
\item{infile}{input necdf file}

\item{outfile}{}

\item{start}{start date}

\item{end}{end date}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo seldate}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\examples{

 infile <- "/Users/ecor/workspace_eclipse/MASTERCLASS/EMC15ENERGY/data/e-obs/tg_0.25deg_reg_v11.0.nc4"
 outfile <- "/Users/ecor/workspace_eclipse/MASTERCLASS/EMC15ENERGY/temp/test0.nc4" 
 
 ncs <- cdo.seldate(infile,outfile,start="1997-05-01",end="1997-05-31")
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

