% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.sellonlatbox.R
\name{cdo.sellonlatbox}
\alias{cdo.sellonlatbox}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.sellonlatbox(x, y, ..., dim = c(1, 1), outdir = NULL, outfile = NULL,
  return.raster = TRUE, parallel = FALSE, npar = NA)
}
\arguments{
\item{x}{coarse input netcdf}

\item{outfile}{output netcdf}

\item{fine_nc}{fine  input netcdf}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo remapnn}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\examples{

 x <- '/H01/SHAREDWORK/ACEWATER2/data/db/climate/chirps/data/daily/netcdf_global/p05/chirps-v2.0.2005.days_p05.nc' 
 y <- extent(c(-17.6608, -5.727799, 10.17488, 18.62197))

	out <- cdo.sellonlatbox(x=x,y=y)
	out2 <- cdo.sellonlatbox(x=x,y=y,dim=c(2,2))


}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

