% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_concate}
\alias{cdo_concate}
\title{Concatenate netcdf files}
\usage{
cdo_concate(dt, intermed_dir)
}
\arguments{
\item{dt}{the input tibble object that contains cmip data and files to process}
}
\value{
The path to the netcdf of the concatenated ncdf.
}
\description{
\code{cdo_concate} Sometimes modeling groups save results for a single experiment
in mutiple necdf files. Here we are going to concatenate the files together. This is some
thing good to do using cdo instead of R because sometimes the netcdf files will repeate
years / months in the different netcdf files and CDO can take care of that with the operator copy.
}
