% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_fldmean_area}
\alias{cdo_fldmean_area}
\title{Calculate the area weighted field mean.}
\usage{
cdo_fldmean_area(name, in_nc, area_nc, intermed_dir)
}
\arguments{
\item{name}{A string for the base name of the intermediate netcdf files that will
be used to label the intermediate netcdf files.}

\item{in_nc}{the path for the input netcdf file that is going to be procssed.}

\item{area_nc}{the path for the area netcdf file that will be used as the area weights.}

\item{intermed_dir}{the path of the directory where all of the intermediate netcdf
files will be saved to.}
}
\value{
The path of the netcdf of the area weighted file.
}
\description{
\code{cdo_fldmean_area} Calculate the area weighted field mean for some varaibel.
}
