% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_land_area}
\alias{cdo_land_area}
\title{Calculate the land area per grid cell}
\usage{
cdo_land_area(dt, intermed_dir)
}
\arguments{
\item{dt}{a wide tibble of area and land fraction data to process.}
}
\value{
The path to the netcdf of the land area.
}
\description{
\code{cdo_land_area} Calculate the land area per grid cell from the land fraction
and gird cell area files. The netcdf returned by this function can be used as a land
area weight when taking the weighted mean.
}
