% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_ocean_area}
\alias{cdo_ocean_area}
\title{Calculate the ocean area per grid cell}
\usage{
cdo_ocean_area(dt, intermed_dir)
}
\arguments{
\item{dt}{a wide data.table of area and land fraction data to process.}
}
\value{
The path to the netcdf of the ocean area.
}
\description{
\code{cdo_ocean_area} Calculate the ocean area per grid cell from the land fraction
and gird cell area files. The netcdf returned by this function can be used as a
area weights to calcualte the mean weighted value over the oceans.
}
