% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_sellonlat_fldmean}
\alias{cdo_sellonlat_fldmean}
\title{Select a lat/lon box for a single area in a single file.}
\usage{
cdo_sellonlat_fldmean(
  basename,
  info,
  in_nc,
  area_nc,
  latlon_df,
  intermed_dir,
  cleanUP = TRUE
)
}
\arguments{
\item{basename}{The string of CMIP information about the file that is being processed, this string
will be appended onto the intermediate netcdf files.}

\item{info}{A tibble of the CMIP information for the in_nc file that is going to be processed.}

\item{in_nc}{The netcdf file to be processed.}

\item{area_nc}{The netcdf file containing the area weights information}

\item{latlon_df}{The dataframe of the lat and lon boundaries, it must include the following columns,
name, lon1, lon2, lat1, and lat2.}

\item{intermed_dir}{The name of the directory to store the intermediate}

\item{cleanUP}{default set to TRUE to remove the intermediate netcdf files.}
}
\value{
A dataframe of the area weighted field mean at each lat and lon box.
}
\description{
\code{internal_cdo_sellonlat} Select the lat/lon box fomr a single
netcdf file.
}
