% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo_yearmonmean}
\alias{cdo_yearmonmean}
\title{}
\usage{
cdo_yearmonmean(name, in_nc, intermed_dir)
}
\arguments{
\item{name}{a string of infomration that will be appended to the name of the netcdf.}

\item{intermed_dir}{the directory where the resulting netcdf will be stored.}
}
\value{
The path to the netcdf of the annual average.
}
\description{
system2(cdoR::cdo_exe, args = c(paste0("setgridarea,", area_nc), in_nc, areadata_nc), stdout = TRUE, stderr = TRUE)
  system2(cdoR::cdo_exe, args = c('fldmean,weights=TRUE', areadata_nc, out_nc), stdout = TRUE, stderr = TRUE)
}
\details{

}
