% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fxn_checks.R
\name{check_annual_continuous}
\alias{check_annual_continuous}
\title{Make sure that the annual data is continuous.}
\usage{
check_annual_continuous(dt)
}
\arguments{
\item{dt}{A tibble object containg CMIP information for a time series of annual data.}
}
\value{
A data frame of monthly data with the problem column indicating if there is an issue with the number of
months in the time series.
}
\description{
\code{check_annual_continuous} Checks to make sure that annual data is continuous and that there are no gaps in the
annual data.
}
