% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fxn_checks.R
\name{check_annual_dupplicates}
\alias{check_annual_dupplicates}
\title{Check for annual data for duplicates.}
\usage{
check_annual_dupplicates(dt)
}
\arguments{
\item{dt}{A tibble object containg CMIP information for a time series of annual data.}
}
\value{
A data frame of monthly data with the problem column indicating if there is an issue with the number of
months in the time series.
}
\description{
\code{check_annual_dupplicates} Checks to see if the annual data has duplicate years, this is espcially
relevant when modeling groups save output in mulitple netcdf files for the same experiment.
If not then return the time series with the relevant problem code.
}
