% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fxn_checks.R
\name{check_cmip_info}
\alias{check_cmip_info}
\title{Check CMIP info}
\usage{
check_cmip_info(dt)
}
\arguments{
\item{dt}{A tibble object containg CMIP information.}
}
\value{
A true or false if the conditions are met.
}
\description{
\code{check_cmip_info} Check to make sure that there is only data for a sinlge model / variable /
experiment and in a tibble of cmip data.
}
