% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_cdo_fxns.R
\name{extract_mean_results}
\alias{extract_mean_results}
\title{Extract mean results}
\usage{
extract_mean_results(out_nc, info)
}
\arguments{
\item{out_nc}{the output netcdf that has the data to extract.}

\item{info}{a data.table of the netcdf CMIP information.}
}
\value{
A data.table of the processed netcdf results.
}
\description{
\code{extract_mean_results} Extract the results from a netcdf file
that does not have gridded data aka as been averaged or something.
}
