% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{fldmean_area}
\alias{fldmean_area}
\title{Calculate the area weighted field mean without using cdo.}
\usage{
fldmean_area(info, in_nc, area_nc, showMessages = FALSE)
}
\arguments{
\item{info}{The dataframe of CMIP information such as variable / ensemble / domain / ect.}

\item{in_nc}{the path for the input netcdf file that is going to be procssed.}

\item{area_nc}{the path for the area netcdf file that will be used as the area weights.}

\item{showMessages}{default set to FALSE to hide messages}
}
\value{
A tibble of the area weighted mean.
}
\description{
\code{fldmean_area} Calculate the area weighted field mean for some varaible, without using
cdo operators. It looks like there might be some issues with how the cdo code is working
with the cmip6 netcdfs, untill those issuse are resolved the workaround will be to use
R to do the area weighting average.
}
