% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_cdo_fxns.R
\name{format_time}
\alias{format_time}
\title{format time}
\usage{
format_time(nc)
}
\arguments{
\item{nc}{an open netcdf file that contains the time information.}
}
\value{
A data.table of the datetime, year, and month.
}
\description{
\code{format_time} Format the time information.
}
