% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProblemData.R
\name{importProblemData}
\alias{importProblemData}
\title{This function imports Problem Data as \code{RasterBrick-class} object}
\usage{
importProblemData(..., files = NULL)
}
\arguments{
\item{...}{further arguments}

\item{files}{netcdf files. If it is \code{NULL} , files are created with \code{\link{cdo.remapnn.conf}}.}
}
\value{
a list of \code{RasterBrick-class} objects
}
\description{
This function imports Problem Data as \code{RasterBrick-class} object
}
\examples{


files <- cdo.remapnn.conf()
maps <- importProblemData(files=files)                                                                                        

}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{brick}},\code{\link{cdo.remapnn.conf}}
}

