% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{internal_cdo_sellonlat}
\alias{internal_cdo_sellonlat}
\title{Select a lat/lon box for a single area in a single file.}
\usage{
internal_cdo_sellonlat(name, box_name, cdo_arg, nc_in, intermed_dir)
}
\arguments{
\item{name}{Some string that will be appended to the netcdf file.}

\item{box_name}{The name of the lat/lon box that is being selected from the netcdf file.}

\item{cdo_arg}{The sellonlat cdo argument created by \code{internal_format_sellonlat_cdo_argument}}

\item{nc_in}{The netcdf file to process}

\item{intermed_dir}{The name of the directory to store the intermediate}
}
\value{
A vector that can be used as the argument to the cdo sellonlat operator.
}
\description{
\code{internal_cdo_sellonlat} Select the lat/lon box fomr a single
netcdf file.
}
