% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{internal_format_coord_input}
\alias{internal_format_coord_input}
\title{Format the data frame of the lon/lat boundaries such that the cutoff values
are relvant to the netcdf file that is being porcessed. This is helpful because
not all of the models report output on the using 0 to 360 and -90 to 90 coords.}
\usage{
internal_format_coord_input(file, input)
}
\arguments{
\item{file}{The netcdf file with the coordinate system the lat/lon coordinate data frame must match.}

\item{input}{A data frame of lat/lon coordinates;}
}
\value{
A tibble of the appropriate lat/lon coordinates.
}
\description{
\code{internal_format_coord_input} Format the lat and lon box boundries to reflect
the coordinate system used by the data netcdf file.
}
