% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{internal_format_sellonlat_cdo_argument}
\alias{internal_format_sellonlat_cdo_argument}
\title{Format sellonlat cdo argument}
\usage{
internal_format_sellonlat_cdo_argument(lon1, lon2, lat1, lat2)
}
\arguments{
\item{lon1}{The western longitude boundary}

\item{lon2}{The eastern longitude boundary}

\item{lat1}{The northern latitude boundary}

\item{lat2}{The southern latitude boundary}
}
\value{
A vector that can be used as the argument to the cdo sellonlat operator.
}
\description{
\code{internal_format_sellonlat_cdo_argument} Format the lat and lon information into a
vector that matches the format expected by the cdo sellonlat operation.
}
