% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.distgrid.R
\name{cdo.distgrid}
\alias{cdo.distgrid}
\title{R Interface for 'cdo.sellonlatbox' and 'cdo.distgrid'}
\usage{
cdo.distgrid(x, y = NULL, ..., dim = c(1, 1), sellonlatbox = TRUE,
  return.raster = TRUE)
}
\arguments{
\item{x}{input netcdf file for 'cdo.sellonlatbox' and the 'cdo.distgrid'}

\item{y}{extent or \code{Extent*} object indicateing the extent to extract or \code{NULL}(dafault). It works with \code{sellonlatbox==TRUE}.}

\item{...}{further arguments for \code{\link{cdo.sellonlatbox}}}

\item{dim}{integer vector reporting the number of rows and columuns of the matrix of tiles into which the cropped map is spit.}

\item{sellonlatbox}{logical value. if it is \code{TRUE} (default) , \code{x} is cropped through \code{\link{cdo.sellonlatbox}}}

\item{return.raster}{logical value. If is \code{TRUE}, outputs are returned as \code{\link{RasterStack-class}} objects.}
}
\description{
R Interface for 'cdo.sellonlatbox' and 'cdo.distgrid'
}
\examples{



\dontrun{

 ncname <- "chirps-v2.0.2005.days_p05.nc"
 url <- "ftp://ftp.chg.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/global_daily/netcdf/p05"
 temp <- tempdir()
 x <- paste(temp,ncname,sep="/")
 download.file(url=paste(url,ncname,sep="/"),destfile=x)




### Africa Extent: extent      : -18.1625, 54.5375, -34.8375, 37.5625  (xmin, xmax, ymin, ymax)
###y <- extent(c(-17.6608, -5.727799, 10.17488, 18.62197)) ## Extent senegal
 y <- extent(c(-18.1625,54.5375,-34.8375,37.5625))

 prec_afr <- cdo.distgrid(x=x,y=y,dim=c(10,10),parallel=TRUE)	


}
}
\seealso{
\code{\link{cdo.sellonlatbox}}
}
