% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.seldate.R
\name{cdo.seldate}
\alias{cdo.seldate}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.seldate(infile, outfile = NULL, start = "1997-05-01",
  end = "1997-05-31", return.raster = TRUE, ...)
}
\arguments{
\item{infile}{input necdf file}

\item{outfile}{}

\item{start}{start date}

\item{end}{end date}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo seldate}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

