% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.seldate.R
\name{cdo.seldate}
\alias{cdo.seldate}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.seldate(infile, outfile = NULL, start = "1997-05-01",
  end = "1997-05-31", return.raster = TRUE, ...)
}
\arguments{
\item{infile}{input necdf file}

\item{outfile}{name of the outputfile}

\item{start}{start date}

\item{end}{end date}

\item{return.raster}{logical value. If is \code{TRUE}, outputs are returned as \code{\link{RasterStack-class}} objects.}

\item{...}{further arguments}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo seldate}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\examples{
 ##

 \dontrun{

 ncname <- "chirps-v2.0.2005.days_p05.nc"
 url <- "ftp://ftp.chg.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/global_daily/netcdf/p05"
 temp <- tempdir()
 x <- paste(temp,ncname,sep="/")
 download.file(url=paste(url,ncname,sep="/"),destfile=x)





 infile <- x
 outfile <- NA
 
 ncs <- cdo.seldate(infile,outfile,start="1997-05-01",end="1997-05-31")
 }
}
\seealso{
\code{\link{system}}
}
\author{
Emanuele Cordano
}
