% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.sellonlatbox.R
\name{cdo.sellonlatbox}
\alias{cdo.sellonlatbox}
\title{R Interface for 'cdo.sellonlatbox'}
\usage{
cdo.sellonlatbox(x, y, ..., dim = c(1, 1), outdir = NULL, outfile = NULL,
  return.raster = TRUE, parallel = FALSE, npar = NA)
}
\arguments{
\item{x}{input netcdf file for 'cdo.sellonlatbox'}

\item{y}{extent or \code{Extent*} object indicateing the extent to extract}

\item{...}{further arguments}

\item{dim}{integer vector reporting the number of rows and columuns of the matrix of tiles into which the cropped map is spit.}

\item{outdir}{output directory}

\item{outfile}{output netcdf}

\item{return.raster}{logical value. If is \code{TRUE}, outputs are returned as \code{\link{RasterStack-class}} objects.}

\item{parallel}{logical value. If it is \code{TRUE} function works in a parallel way using \code{doParallel} package.  Default is \code{FALSE}.}

\item{npar}{number of working cores (used if \code{parall==TRUE}). In case of \code{NA} (default) and \code{parallel==TRUE} function uses all but one cores of the CPU. See \code{detectCores}.}
}
\description{
R Interface for 'cdo.sellonlatbox'
}
\note{
This function calls \code{cdo sellonlatbox}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

