% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.sellonlatbox.R
\name{cdo.sellonlatbox}
\alias{cdo.sellonlatbox}
\title{R Interface for 'cdo.seldate'}
\usage{
cdo.sellonlatbox(x, y, ..., dim = c(1, 1), outdir = NULL, outfile = NULL,
  return.raster = TRUE, parallel = FALSE, npar = NA,
  use.distbase = FALSE)
}
\arguments{
\item{x}{input netcdf file for 'cdo.sellonlatbox'}

\item{y}{extent or \code{Extent*} object indicateing the extent to extract}

\item{outdir}{output directory}

\item{outfile}{output netcdf}

\item{return.raster}{logical value. If is \code{TRUE}, outputs are returned as \code{\link{RasterStack-Class}} objects.}

\item{fine_nc}{fine  input netcdf}
}
\description{
R Interface for 'cdo.seldate'
}
\note{
This function calls \code{cdo remapnn}. 
See \url{http://www.fourtythree.org/tech/remapping-a-netcdf-file-using-cdo/},
\url{https://code.zmaw.de/projects/cdo/embedded/1.4.7/cdo.html}
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{system}}
}

